auth_service/requirements.txt
aio-pika==9.2.3
aiormq==6.7.7
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.16
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
mysql-connector-python==8.4.0
orjson==3.11.4
pamqp==3.2.1
passlib==1.7.4
pika==1.3.2
propcache==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
PyJWT==2.8.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.1
rich==14.2.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.20.0
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
